backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
transformers==4.35.2
torch==2.4.0
spacy==3.7.2
requests==2.31.0
beautifulsoup4==4.12.2
pytesseract==0.3.10
pillow==10.3.0
opencv-python-headless==4.8.1.78
chardet==5.2.0
pdf2image==1.16.3
PyPDF2==3.0.1
werkzeug==2.3.7
sentencepiece==0.2.0
protobuf==4.25.3